improved the output

+++ OPAL/ai/src/main/scala/org/opalj/ai/InstructionCountBoundedAI.scala
@@ -94,7 +94,9 @@ object InstructionCountBoundedAI {
         }
 
         if (upperBound > 65535.0 /*Max Length*/ * 10.0) {
-            println(Console.YELLOW+"[warn] evaluation (up to: "+upperBound+" instructions) may take execessively long"+
+            println(Console.YELLOW+
+                "[warn] evaluation (up to: "+upperBound.toInt+
+                " instructions) may take execessively long"+
                 "; instructions size="+code.instructions.size+
                 "; exception handlers="+code.exceptionHandlers.size+
                 "; maxEvaluationFactor="+maxEvaluationFactor + Console.RESET)
